requirements.txt
asgiref==3.8.1
chardet==5.2.0
dj-database-url==2.2.0
Django==5.0.6
gunicorn==22.0.0
packaging==24.1
pillow==10.3.0
psycopg==3.1.19
reportlab==4.2.0
sqlparse==0.5.0
typing_extensions==4.12.2
